Ignore non notebook files when processing arguments on the command line
Assume to have a directory with file1.ipynb and file2.py. When calling pytest file* nbvalx will only run file1.ipynb, and effectively ignore file2.py. This is consistent with the behavior without input arguments, but the assertion that this commit changed into an if was effectively halting pytest with an assertion error